FindResult -> VideoReference
